requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pamqp==3.3.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.179
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
